Fix getXsltProcessor()
The method used to create a new instance of self, which is wrong
because that class might be specific to the XML document and not
applicable to XSLTs.  Now it creates an instance of
alcamo\dom\Document.

For the user it makes no difference because the XSLT document object
is not accesible from outside of getXsltProcessor().

	modified:   src/Document.php